리팩토링: 1단계 코드리뷰 반영
1. 불필요한 인터페이스와 클래스 제거
현 단계에서 불필요한 UserRepsitory 인터페이스를 삭제하였으며,
동일한 역할을 하는 UserServicer와 MemoryUserRepository를  UserRepsitory
클래스로 통합하였습니다.

2. 회원가입시 중복검사를 별도의 메서드가 아닌 ifPresentOrElse를 통해
   한번에 처리하도록 하였습니다.

3. 코드리뷰를 받은것은 아니나 UserController에서 해당 유저를 조회할때
   유저가 존재하는지 여부를 먼저 확인하도록 변경하였습니다.
