--- src/main/java/com/kakao/cafe/controller/HomeController.java
@@ -1,13 +1,0 @@
-package com.kakao.cafe.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-
-	@GetMapping("/")
-	public String index() {
-		return "index";
-	}
-}
