+++ src/main/java/com/kakao/cafe/controller/UserController.java
@@ -9,20 +9,20 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
-import com.kakao.cafe.Service.UserService;
 import com.kakao.cafe.domain.User;
+import com.kakao.cafe.repository.UserRepository;
 
 @Controller
 public class UserController {
-
-	private final UserService userService;
+	private static final String ERROR_MESSAGE_NO_SUCH_USER_ID = "존재하지 않는 user ID 입니다.";
+	private final UserRepository userRepository;
 
 	@Autowired
-	public UserController(UserService userService) {
-		this.userService = userService;
+	public UserController(UserRepository userRepository) {
+		this.userRepository = userRepository;
 	}
 
-	@GetMapping("/user/join")
+	@GetMapping("/user/new")
 	public String createForm() {
 		return "user/form";
 	}
@@ -30,13 +30,13 @@ public class UserController {
 	@PostMapping("/users")
 	public String create(String userId, String password, String name, String email) {
 		User user = new User(userId, password, name, email);
-		userService.join(user);
+		userRepository.save(user);
 		return "redirect:/users";
 	}
 
 	@GetMapping("/users")
 	public String users(Model model) {
-		List<User> users = userService.findUsers();
+		List<User> users = userRepository.findAll();
 		model.addAttribute("users", users);
 		return "user/list";
 	}
@@ -48,7 +48,8 @@ public class UserController {
 
 	@GetMapping("/users/{userId}")
 	public String profile(@PathVariable String userId, Model model) {
-		model.addAttribute("user", userService.findOne(userId).get());
+		User user = userRepository.findByUserId(userId).orElseThrow(() -> new IllegalStateException(ERROR_MESSAGE_NO_SUCH_USER_ID));
+		model.addAttribute("user", user);
 		return "user/profile";
 	}
 }
