+++ src/main/java/com/kakao/cafe/repository/UserRepository.java
@@ -1,13 +1,31 @@
 package com.kakao.cafe.repository;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.stereotype.Repository;
+
 import com.kakao.cafe.domain.User;
 
-public interface UserRepository {
+@Repository
+public class UserRepository {
+	private static final String ERROR_MESSAGE_DUPLICATED_USER_ID = "이미 존재하는 id 입니다.";
+	private static final List<User> userList = new ArrayList<>();
 
-	void save(User user);
-	Optional<User> findByUserId(String userId);
-	List<User> findAll();
+	public void save(User user) {
+		findByUserId(user.getUserId())
+			.ifPresentOrElse(u -> {throw new IllegalStateException(ERROR_MESSAGE_DUPLICATED_USER_ID);},
+				() -> userList.add(user));
+	}
+
+	public List<User> findAll() {
+		return new ArrayList<>(List.copyOf(userList));
+	}
+
+	public Optional<User> findByUserId(String userId) {
+		return userList.stream()
+			.filter(user -> user.getUserId().equals(userId))
+			.findAny();
+	}
 }
